requirements.txt
bcrypt==4.0.1
cffi==1.15.1
cryptography==41.0.3
et-xmlfile==1.1.0
numpy==1.25.2
openpyxl==3.1.2
pandas==2.1.0
paramiko==3.3.1
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
python-nmap==0.7.1
pytz==2023.3.post1
six==1.16.0
timeout-decorator==0.5.0
tqdm==4.66.1
tzdata==2023.3
